+++ app/src/main/java/com/example/bazar1177s/network/ProductService.kt
@@ -1,7 +1,6 @@
 package com.example.bazar1177s.network
 
 import com.example.bazar1177s.model.*
-import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
@@ -12,24 +11,24 @@ interface ProductService {
 //auth
 
     @POST("/api/user/register")
-    suspend fun loginAndGetAccessToken(@Body user:LoginUser):LoginResponse
+    suspend fun loginAndGetAccessToken(@Body user: LoginUser): LoginResponse
 
     @GET("/api/shops")
-    suspend fun getShops():List<Shop>
+    suspend fun getShops(): List<Shop>
 
-    @GET( "api/categoriesByShopId/{id}")
-    suspend fun getCategoriesByShops(@Path("id") id: Int):BaseResponseList<ShopCategory>
+    @GET("api/categoriesByShopId/{id}")
+    suspend fun getCategoriesByShops(@Path("id") id: Int): BaseResponseList<ShopCategory>
 
     @GET("api/productsByCategoryId/{id}")
-    suspend fun getProductsByCategoryId(@Path("id") id: Int):BaseResponseList<Product>
+    suspend fun getProductsByCategoryId(@Path("id") id: Int): BaseResponseList<Product>
 
     @GET("api/carousel")
-    suspend fun getCarouselProducts():BaseResponseList<CarouselProduct>
+    suspend fun getCarouselProducts(): BaseResponseList<CarouselProduct>
 
     @GET("api/discounts")
-    suspend fun getDiscountProducts():BaseResponseList<DiscountProduct>
+    suspend fun getDiscountProducts(): BaseResponseList<DiscountProduct>
 
     @GET("api/product/{id}")
-    suspend fun getProduct(@Path("id") id: Int):BaseResponseObject<Product>
+    suspend fun getProduct(@Path("id") id: Long): BaseResponseObject<Product>
 
 }
